[jit] split TestLogging, TestDict, TestList
